refactor: split hooks into separate files
